inventory_llm/requirements.txt
langchain
langchain-community
langchain-groq
unstructured
selenium
chromadb
streamlit
pandas
python-dotenv
tiktoken
faiss-cpu
protobuf
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
